refactor: move contact models to appropriate directory
