+++ src/core/coreav.cpp
@@ -21,8 +21,8 @@
 #include "coreav.h"
 #include "core.h"
 #include "src/audio/audio.h"
-#include "src/model/friend.h"
-#include "src/model/group.h"
+#include "src/model/contacts/friend.h"
+#include "src/model/contacts/group.h"
 #include "src/persistence/settings.h"
 #include "src/video/corevideosource.h"
 #include "src/video/videoframe.h"
