+++ src/model/contacts/friend.h
