+++ src/persistence/offlinemsgengine.cpp
@@ -19,7 +19,7 @@
 
 #include "offlinemsgengine.h"
 #include "src/core/core.h"
-#include "src/model/friend.h"
+#include "src/model/contacts/friend.h"
 #include "src/nexus.h"
 #include "src/persistence/profile.h"
 #include "src/persistence/settings.h"
