+++ src/video/netcamview.cpp
@@ -20,7 +20,7 @@
 #include "netcamview.h"
 #include "camerasource.h"
 #include "src/core/core.h"
-#include "src/model/friend.h"
+#include "src/model/contacts/friend.h"
 #include "src/friendlist.h"
 #include "src/nexus.h"
 #include "src/persistence/profile.h"
