+++ src/widget/form/groupchatform.cpp
@@ -22,9 +22,9 @@
 #include "tabcompleter.h"
 #include "src/core/core.h"
 #include "src/core/coreav.h"
-#include "src/model/friend.h"
+#include "src/model/contacts/friend.h"
 #include "src/friendlist.h"
-#include "src/model/group.h"
+#include "src/model/contacts/group.h"
 #include "src/video/groupnetcamview.h"
 #include "src/widget/chatformheader.h"
 #include "src/widget/flowlayout.h"
