+++ src/widget/friendwidget.cpp
@@ -24,10 +24,10 @@
 #include "maskablepixmapwidget.h"
 
 #include "src/core/core.h"
-#include "src/model/friend.h"
+#include "src/model/contacts/friend.h"
 #include "src/model/about/aboutfriend.h"
 #include "src/friendlist.h"
-#include "src/model/group.h"
+#include "src/model/contacts/group.h"
 #include "src/grouplist.h"
 #include "src/persistence/settings.h"
 #include "src/widget/about/aboutfriendform.h"
